requirements.txt
streamlit
torch
torchvision
numpy
pandas
matplotlib
Pillow
lime
scikit-image
plotly
captum
fairlearn
scikit-learn
